app: Extracted common UI logic.
Renamed `ServicesUi` to `ServicesOverviewUi`, since it renders the UI
for the services overview.

In `ServicesUi` now the functionality is contained to dispatch the UI
state, and to render the loading and error screens.
